- Tried to reduce the number of compiler warnings generated.
- The cross compiler GNUmakefile.os4 variant now also logs all warnings
  and errors to a compiler.log file.


git-svn-id: file:///Users/olsen/Code/migration-svn-zu-git/logical-line-staging/clib2/trunk@15055 87f5fb63-7c3d-0410-a384-fd976d0f7a62

+++ library/math_lround.c
@@ -1,5 +1,5 @@
 /*
- * $Id: math_lround.c,v 1.2 2005-10-09 10:38:55 obarthel Exp $
+ * $Id: math_lround.c,v 1.3 2005-10-27 08:58:41 obarthel Exp $
  *
  * :ts=4
  *
@@ -82,7 +82,7 @@ lround(double x)
           result = msw >> (20 - exponent_less_1023);
         }
     }
-  else if (exponent_less_1023 < (8 * sizeof (long int)) - 1)
+  else if (exponent_less_1023 < (LONG)(8 * sizeof (long int)) - 1)
     {
       if (exponent_less_1023 >= 52)
         result = ((long int) msw << (exponent_less_1023 - 20)) | (lsw << (exponent_less_1023 - 52));
